Move modifyConfigJson and supporting functions to new file
	modified:   staging/src/k8s.io/kubectl/pkg/cmd/config/set.go
	new file:   staging/src/k8s.io/kubectl/pkg/cmd/config/set_jsonpath.go
